devfreq: Fix type of frequency table
Frequencies are always expressed as ulong, but devfreq's table is of
type uint *. Fix the mismatch.

Reviewed-on: http://git-master/r/352720

+++ include/linux/devfreq.h
@@ -49,7 +49,7 @@ struct devfreq_dev_profile {
 	int (*get_cur_freq)(struct device *dev, unsigned long *freq);
 	void (*exit)(struct device *dev);
 
-	unsigned int *freq_table;
+	unsigned long *freq_table;
 	unsigned int max_state;
 	const struct devfreq_governor_data *governor_data;
 	unsigned int num_governor_data;
